chore: add .gitignore file to exclude unnecessary files
This commit adds a .gitignore file to the project. The file
excludes files such as .env, .DS_Store, and other files that
should not be tracked by Git. This helps to keep the repository
clean and prevents sensitive information from being committed.
The .gitignore file was generated using the toptal.com service
and includes templates for dotenv, direnv, terraform, and macOS.